HandAndBrushMover and ToothInfo
Move hand & brush with lerp and animation curves for pos and rot, the rot is based on the nearest keyframe transforms defined. Then the brush action is a couple of lerps for up and down, with animation curves also. The tooth has info on letters needed and dirtiness, but they are not assigned to by anything.