Fixes for v1.0:
rigged score to title scene and end scene. Fixed score order. Fixed scores acting weird when starting a new game. Fixed error at the end of level that didn't end the game and you had to wait for timer to reach 0 (added a new scene before the first scene with a game manager).